+++ crates/wdl-lint/src/rules/deprecated_placeholder.rs
@@ -101,7 +101,11 @@ impl Rule for DeprecatedPlaceholderRule {
     }
 
     fn related_rules(&self) -> &[&'static str] {
-        &["DeprecatedObject", "ExpectedRuntimeKeys"]
+        &[
+            "DeprecatedObject",
+            "ExpectedRuntimeKeys",
+            "OptionalInputSafety",
+        ]
     }
 }