+++ crates/wdl-lint/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+* New lint rule `OptionalInputSafety` to warn when optional values appear in task command placeholders without an explicit `None` guard ([#715](https://github.com/stjude-rust-labs/sprocket/pull/715)).
 * New lint rule `ParameterDescription` to ensure parameters and outputs have proper descriptions ([#449](https://github.com/stjude-rust-labs/sprocket/pull/449)).
 
 #### Removed